+++ apps/native-component-list/src/screens/Audio/expo-audio/AudioScreen.tsx
@@ -4,6 +4,7 @@ import { PixelRatio, ScrollView, StyleSheet } from 'react-native';
 
 import AudioModeSelector from './AudioModeSelector';
 import AudioPlayer from './AudioPlayer';
+import AudioQueuePlayer from './AudioQueuePlayer';
 import HeadingText from '../../../components/HeadingText';
 import ListButton from '../../../components/ListButton';
 
@@ -31,6 +32,16 @@ export default function AudioScreen(props: any) {
       <HeadingText>Local asset player</HeadingText>
       <AudioPlayer
         source={require('../../../../assets/sounds/polonez.mp3')}
+        style={styles.player}
+      />
+      <HeadingText>Queue player</HeadingText>
+      <AudioQueuePlayer
+        source={[
+          {
+            uri: 'https://p.scdn.co/mp3-preview/f7a8ab9c5768009b65a30e9162555e8f21046f46?cid=162b7dc01f3a4a2ca32ed3cec83d1e02',
+          },
+          require('../../../../assets/sounds/polonez.mp3'),
+        ]}
         style={styles.player}
       />
     </ScrollView>